refactor(infra,cli): Move EnsureTskPath to FlatFileRepositoryFactory
