+++ tests/Tsk.Infrastructure.Tests/Utils/HelpersTests.ExpandUserPath.cs
@@ -1,4 +1,4 @@
-using Tsk.CLI.Utils;
+using Tsk.Infrastructure.Utils;
 using Tsk.TestSupport;
 
 namespace Tsk.CLI.Tests.Utils;
